+++ CHANGELOG.md
@@ -1,0 +1,15 @@
+# Portfolio Changelog
+All notable changes to this project will be documented in this file. Dates are displayed in UTC +8.
+
+
+### v1.0.0
+> August 25, 2023
+- Added header with responsive navigator section on it.
+- Logo no longer blurs because I learned from mistakes. Sheesh progress 😅
+- Added hero section with the improved 2021 design concept.
+- Added featured projects section.
+- Added 6 project on projects section.
+- Added about section.
+- Added contact section.
+- Added footer section.
+- Responsive layout, support large to small screen sizes.
