+++ dev-CHANGELOG.md
@@ -1,0 +1,14 @@
+# Portfolio Development Changelog
+All notable development changes to this project will be documented in this file. Dates are displayed in UTC +8.
+
+
+### v23.8.25b
+> August 25, 2023
+- Fixed navigation menu bug on screen width change.
+- Separate files/components to its right folder.
+- Installed framer motion library `https://www.npmjs.com/package/framer-motion`.
+- Partially overhaul header section.
+
+### v23.0.0
+> Month 0, 2023
+<!-- Old dev changes were not recorded 😞 -->
