+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "emailjs-com": "^3.2.0",
+    "framer-motion": "^10.16.1",
     "prettier-plugin-tailwindcss": "^0.5.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
