--- src/assets/svg/FacebookIcon.js
@@ -1,19 +1,0 @@
-function FacebookIcon({ className, stroke }) {
-  return (
-    <svg className={className} viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_858_86)">
-        <path
-          d="M8.65878 9.00008V15.5H6.36908V9.00132V8.50132H5.86908H3.69995V6.60625H5.87173H6.37173V6.10625V3.897C6.37173 2.74878 6.72816 1.91638 7.30254 1.36905C7.88055 0.818251 8.73832 0.5 9.84508 0.5C10.3957 0.5 10.9645 0.546895 11.3999 0.594714C11.5098 0.606793 11.6107 0.618866 11.7 0.630149V2.15981H10.8733C10.1193 2.15981 9.54044 2.38634 9.15669 2.8095C8.78169 3.22302 8.66007 3.75104 8.66007 4.22558V6.10502V6.60502H9.16007H11.4896L11.1849 8.50008H9.15878H8.65878V9.00008Z"
-          stroke={stroke}
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_858_86">
-          <rect width="16" height="16" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
-
-export default FacebookIcon;
