--- src/assets/svg/LinkedinIcon.js
@@ -1,19 +1,0 @@
-function LinkedinIcon({ className, stroke }) {
-  return (
-    <svg className={className} viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_858_86)">
-        <path
-          d="M6.01337 14.8245C6.01442 14.5742 6.01577 14.2405 6.01724 13.8472C6.0206 12.9535 6.02463 11.7517 6.02732 10.5195C6.03162 8.54856 6.0326 6.48105 6.02133 5.48462H8.44595V6.30103L8.00988 6.98301H8.92306H8.94595H9.44595V6.59837C9.64863 6.30213 9.90603 5.99035 10.2676 5.74015C10.6716 5.46069 11.232 5.23989 12.0443 5.23989C13.0636 5.23989 13.9085 5.56982 14.5024 6.21433C15.1001 6.86287 15.5 7.8913 15.5 9.39392V14.8245H13.0639V9.79321C13.0639 9.0524 12.9341 8.35079 12.5734 7.82096C12.1909 7.25913 11.5927 6.95475 10.8236 6.95475C9.60387 6.95475 8.89345 7.78726 8.59753 8.52232C8.46539 8.84233 8.44595 9.23908 8.44595 9.5499V14.8245H6.01337ZM3.13931 5.48462V14.8245H0.703174V5.48462H3.13931ZM3.36545 1.7913C3.36416 2.15714 3.22809 2.46888 2.99684 2.69065C2.76477 2.91321 2.40797 3.0721 1.92194 3.0721H1.89906C1.00981 3.0721 0.5 2.49131 0.5 1.78605C0.5 1.41043 0.640106 1.09771 0.874337 0.876813C1.10949 0.655045 1.46789 0.5 1.94483 0.5C2.42062 0.5 2.76516 0.65378 2.99058 0.87187C3.2159 1.08986 3.35605 1.40459 3.36545 1.7913Z"
-          stroke={stroke}
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_858_86">
-          <rect width="16" height="16" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
-
-export default LinkedinIcon;
