--- src/assets/svg/NavigationIcon.js
@@ -1,12 +1,0 @@
-function NavigationIcon({ className, fill }) {
-  return (
-    <svg className={className} viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M2 4.5C2 4.22386 2.22386 4 2.5 4H17.5C17.7761 4 18 4.22386 18 4.5C18 4.77614 17.7761 5 17.5 5H2.5C2.22386 5 2 4.77614 2 4.5ZM2 9.5C2 9.22386 2.22386 9 2.5 9H17.5C17.7761 9 18 9.22386 18 9.5C18 9.77614 17.7761 10 17.5 10H2.5C2.22386 10 2 9.77614 2 9.5ZM2.5 14C2.22386 14 2 14.2239 2 14.5C2 14.7761 2.22386 15 2.5 15H17.5C17.7761 15 18 14.7761 18 14.5C18 14.2239 17.7761 14 17.5 14H2.5Z"
-        fill={fill}
-      />
-    </svg>
-  );
-}
-
-export default NavigationIcon;
