+++ src/assets/svg/RounderBG.js
@@ -1,0 +1,9 @@
+function RounderBG({ className }) {
+  return (
+    <svg className={className} viewBox="0 0 2560 158" xmlns="http://www.w3.org/2000/svg">
+      <path d="M0 0L0 157.453H2560V5.2187e-06C2222.3 97.1869 1773.09 156.453 1280 156.453C786.914 156.453 337.698 97.1869 0 0Z" />
+    </svg>
+  );
+}
+
+export default RounderBG;
