+++ src/components/icons/EmailIcon.js
@@ -1,0 +1,12 @@
+function EmailIcon({ className, fill }) {
+  return (
+    <svg className={className} viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M13.5 3C14.8807 3 16 4.11929 16 5.5V13.5C16 14.8807 14.8807 16 13.5 16H2.5C1.11929 16 0 14.8807 0 13.5V5.5C0 4.11929 1.11929 3 2.5 3H13.5ZM15 6.961L8.25351 10.931C8.12311 11.0077 7.96661 11.0205 7.82751 10.9693L7.74649 10.931L1 6.963V13.5C1 14.3284 1.67157 15 2.5 15H13.5C14.3284 15 15 14.3284 15 13.5V6.961ZM13.5 4H2.5C1.67157 4 1 4.67157 1 5.5V5.802L8 9.91991L15 5.801V5.5C15 4.67157 14.3284 4 13.5 4Z"
+        fill={fill}
+      />
+    </svg>
+  );
+}
+
+export default EmailIcon;
