+++ src/components/icons/LinkedinIcon.js
@@ -1,0 +1,33 @@
+function LinkedinIcon({ className, fill }) {
+  return (
+    <svg className={className} viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
+      <g clipPath="url(#clip0_858_86)">
+        <path
+          fillRule="evenodd"
+          clip-rule="evenodd"
+          d="M5.51121 5.98462C5.51791 6.136 5.52252 6.49183 5.52553 6.98462C5.53792 9.01258 5.52317 13.3601 5.5154 15.3245C5.51296 15.9424 5.51121 16.3245 5.51121 16.3245H8.94595V10.5499C8.94595 10.2408 8.96883 9.93167 9.06046 9.71126C9.30801 9.09445 9.87333 8.45475 10.8236 8.45475C12.0672 8.45475 12.5639 9.40216 12.5639 10.7932V16.3245H16V10.3939C16 7.21681 14.3055 5.73989 12.0443 5.73989C11.1094 5.73989 10.439 6.00331 9.94595 6.35512C9.47747 6.68939 9.16908 7.10347 8.94595 7.44722V5.98462H5.51121ZM8.94595 7.44722V7.48301H8.92306L8.94595 7.44722ZM7.94595 6.98462H6.52547C6.53231 8.0914 6.53097 9.84772 6.52732 11.5206C6.52463 12.7531 6.5206 13.9552 6.51724 14.8491L6.51541 15.3245H7.94595V10.5499C7.94595 10.353 7.95232 10.0883 8 9.8147V7H7.94595V6.98462ZM10.1237 7.53647C10.3394 7.4839 10.5725 7.45475 10.8236 7.45475C11.7401 7.45475 12.5014 7.82666 12.9867 8.53959C13.4285 9.18852 13.5639 10.0071 13.5639 10.7932V15.3245H15V10.3939C15 8.97724 14.6239 8.08389 14.1348 7.55318C13.6531 7.03052 12.9522 6.73989 12.0443 6.73989C11.3312 6.73989 10.8705 6.93104 10.5521 7.15133C10.3878 7.26499 10.2476 7.39596 10.1237 7.53647Z"
+          fill={fill}
+        />
+        <path
+          fillRule="evenodd"
+          clip-rule="evenodd"
+          d="M3.86546 2.78605C3.86546 3.77925 3.11985 4.5721 1.92194 4.5721H1.89906C0.745612 4.5721 0 3.77925 0 2.78605C0 1.7728 0.768497 1 1.94483 1C3.12125 1 3.84397 1.77138 3.86546 2.78605ZM2.65076 3.32978C2.77748 3.20825 2.86296 3.03087 2.86541 2.79639C2.85741 2.53127 2.76356 2.34794 2.64291 2.23122C2.5225 2.11471 2.3082 2 1.94483 2C1.57912 2 1.34853 2.11689 1.21738 2.24057C1.08809 2.3625 1 2.54143 1 2.78605C1 3.20338 1.27401 3.5721 1.89906 3.5721H1.92194C2.29504 3.5721 2.52276 3.45253 2.65076 3.32978Z"
+          fill={fill}
+        />
+        <path
+          fillRule="evenodd"
+          clip-rule="evenodd"
+          d="M3.63931 5.98462H0.203174V16.3245H3.63931V5.98462ZM2.63931 15.3245V6.98462H1.20317V15.3245H2.63931Z"
+          fill={fill}
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_858_86">
+          <rect width="16" height="16" fill={fill} />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export default LinkedinIcon;
