+++ src/components/icons/SendIcon.js
@@ -1,0 +1,12 @@
+function SendIcon({ className, fill }) {
+  return (
+    <svg className={className} viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M2.18412 2.11244C2.33657 1.98818 2.54771 1.96483 2.72363 2.05279L17.7236 9.55279C17.893 9.63749 18 9.81062 18 10C18 10.1894 17.893 10.3625 17.7236 10.4472L2.72363 17.9472C2.54771 18.0352 2.33657 18.0118 2.18412 17.8876C2.03167 17.7633 1.96623 17.5612 2.0169 17.3712L3.98255 10L2.0169 2.62884C1.96623 2.4388 2.03167 2.2367 2.18412 2.11244ZM4.88416 10.5L3.26911 16.5564L16.382 10L3.26911 3.44357L4.88416 9.5H11.5C11.7762 9.5 12 9.72386 12 10C12 10.2761 11.7762 10.5 11.5 10.5H4.88416Z"
+        fill={fill}
+      />
+    </svg>
+  );
+}
+
+export default SendIcon;
